Renamed toggle to a more descriptive name for the widget toggle
